refactor(test): P5-1 follow-up — extract Wave 5 css.5 audit cluster
Continue test_widget_bridge.cpp split. Extracts the 1,599-line Wave 5
css.5 audit (the 49 entries PR #1649 flipped from partial/DIVERGE to
supported) into a sibling TU.

Per-entry runtime-path coverage: JS shim → bridge → View slot
round-trip for backgroundPosition / backgroundSize / textShadow /
border + side longhands + borderRadius + per-corner radius / boxShadow
parser (dx/dy/blur/spread/color/inset) / opacity / outline shorthand
+ longhands / textOverflow / transformOrigin (keyword + percentage)
/ zIndex / backdropFilter blur / display flex-fallback / overflow +
per-axis / many more. Categories per planning/WAVE5-CSS-AUDIT.md
(Cat-1 genuinely supported / Cat-2 architectural caveat / Cat-3
newly-wired in this batch).

* test_widget_bridge.cpp: 9,074 → 7,475 lines (-1,599, -17.6%)
* New target pulp-test-widget-bridge-wave5-css: 1,639 lines, 45
  TEST_CASEs, 218 assertions.

Single biggest single-file reduction this session. No behaviour
change.
