+++ test/test_widget_bridge.cpp
@@ -6764,1605 +6764,6 @@ TEST_CASE("CSSStyleDeclaration mixBlendMode plus-lighter -> kPlus",
 }
 
 
-// ──────────────────────────────────────────────────────────────────────
-// Wave 5 css.5 — audit of the 49 entries flipped by PR #1649 from
-// `partial`/DIVERGE to `supported`. These tests exercise the *runtime*
-// path (JS shim → bridge → View slot) for each catalog claim, so a
-// future drift between catalog metadata and shipped behavior surfaces
-// as a CI failure rather than silent paper coverage.
-//
-// Categories per planning/WAVE5-CSS-AUDIT.md:
-//   • Cat-1 — genuinely supported (regression test below proves it).
-//   • Cat-2 — architectural caveat (test proves the documented
-//     no-crash + sensible-fallback contract; the catalog `notes`
-//     field cites the Pulp design constraint that justifies it).
-//   • Cat-3 — was unwired by #1649; now wired in this PR. Test proves
-//     the new bridge fn's round-trip.
-// ──────────────────────────────────────────────────────────────────────
-
-// Cat-3 — backgroundPosition / backgroundSize were referenced from
-// web-compat-style-decl.js inside `typeof set... === "function"` guards
-// but no bridge fn was registered. PR #1649 declared them `supported`
-// without wiring; Wave 5 css.5 lands the registration so the
-// round-trip is honest.
-TEST_CASE("Wave5 css/backgroundPosition wires JS → bridge → View slot",
-          "[view][bridge][css][wave5][issue-1649]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.backgroundPosition = 'center';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE(p->background_position() == "center");
-
-    // Direct bridge call also round-trips (covers React's prop-applier path).
-    bridge.load_script("setBackgroundPosition('p', 'top left')");
-    REQUIRE(p->background_position() == "top left");
-
-    bridge.load_script("setBackgroundPosition('p', '50% 50%')");
-    REQUIRE(p->background_position() == "50% 50%");
-
-    bridge.load_script("setBackgroundPosition('p', '10px 20px')");
-    REQUIRE(p->background_position() == "10px 20px");
-}
-
-TEST_CASE("Wave5 css/backgroundSize wires JS → bridge → View slot",
-          "[view][bridge][css][wave5][issue-1649]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.backgroundSize = 'cover';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE(p->background_size() == "cover");
-
-    bridge.load_script("setBackgroundSize('p', 'contain')");
-    REQUIRE(p->background_size() == "contain");
-
-    bridge.load_script("setBackgroundSize('p', 'auto')");
-    REQUIRE(p->background_size() == "auto");
-
-    bridge.load_script("setBackgroundSize('p', '100px 200px')");
-    REQUIRE(p->background_size() == "100px 200px");
-
-    bridge.load_script("setBackgroundSize('p', '50% 75%')");
-    REQUIRE(p->background_size() == "50% 75%");
-}
-
-// Cat-3 — textShadow CSS shorthand. The shim parses
-// `<dx>px <dy>px <blur>px <color>` and calls setTextShadow(); the
-// bridge fn was unregistered before Wave 5 css.5. The new fn fans out
-// into the existing 3 per-attribute slots so React's setTextShadow*
-// props keep working unchanged.
-TEST_CASE("Wave5 css/textShadow CSS shorthand fans into per-attribute slots",
-          "[view][bridge][css][wave5][issue-1649]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.textShadow = '2px 3px 4px rgba(0,0,0,0.5)';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    // Shim parses dx=2 dy=3 blur=4 + color → all 3 slots populated.
-    // text_shadow_color stores the literal CSS-color string the shim
-    // resolved (parseCSSColor returns "#rrggbb" or the original token).
-    REQUIRE_THAT(p->text_shadow_offset_x(), WithinAbs(2.0f, 1e-5f));
-    REQUIRE_THAT(p->text_shadow_offset_y(), WithinAbs(3.0f, 1e-5f));
-    REQUIRE_THAT(p->text_shadow_radius(), WithinAbs(4.0f, 1e-5f));
-    REQUIRE(!p->text_shadow_color().empty());
-
-    // Direct bridge call (mirrors what the shim emits):
-    bridge.load_script("setTextShadow('p', 5, 6, 7, '#ff0080')");
-    REQUIRE_THAT(p->text_shadow_offset_x(), WithinAbs(5.0f, 1e-5f));
-    REQUIRE_THAT(p->text_shadow_offset_y(), WithinAbs(6.0f, 1e-5f));
-    REQUIRE_THAT(p->text_shadow_radius(), WithinAbs(7.0f, 1e-5f));
-    REQUIRE(p->text_shadow_color() == "#ff0080");
-}
-
-// ──────────────────────────────────────────────────────────────────────
-// Cat-1 regression coverage — the catalog's `supported` claim is real.
-// Each test exercises the JS shim → bridge → View slot path with a
-// concrete value and asserts the runtime effect.
-// ──────────────────────────────────────────────────────────────────────
-
-TEST_CASE("Wave5 css/border shorthand routes per-attribute (preserves radius)",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.borderRadius = '12px';
-        s.border = '3px solid #ff0000';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE(p->has_border());
-    REQUIRE_THAT(p->border_width(), WithinAbs(3.0f, 1e-5f));
-    REQUIRE(p->border_color().r8() == 0xff);
-    // The Wave 5 audit confirms the per-attribute fix from #1169:
-    // setting `border` shorthand does NOT zero a previously-set radius.
-    REQUIRE_THAT(p->corner_radius(), WithinAbs(12.0f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/borderTop/Right/Bottom/Left shorthand routes per-side",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.borderTop    = '4px solid #ff0000';
-        s.borderRight  = '5px solid #00ff00';
-        s.borderBottom = '6px solid #0000ff';
-        s.borderLeft   = '7px solid #ffff00';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE_THAT(p->border_top_width(), WithinAbs(4.0f, 1e-5f));
-    REQUIRE_THAT(p->border_right_width(), WithinAbs(5.0f, 1e-5f));
-    REQUIRE_THAT(p->border_bottom_width(), WithinAbs(6.0f, 1e-5f));
-    REQUIRE_THAT(p->border_left_width(), WithinAbs(7.0f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/borderRadius accepts px and routes to setBorderRadius",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.borderRadius = '8.5px';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->corner_radius(), WithinAbs(8.5f, 1e-5f));
-
-    // Cat-2 architectural caveat — `%` parses, but the bridge slot is
-    // scalar (no box-relative resolution). We accept the keyword so
-    // the JS layer doesn't crash; the value lands as a px-equivalent
-    // best-effort. Catalog `notes` cite arch-skia-rrect-single-radius.
-    bridge.load_script("var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true }); s2.borderRadius = '10%';");
-    // Should not crash; some non-zero radius landed.
-    REQUIRE(p->corner_radius() >= 0.0f);
-}
-
-TEST_CASE("Wave5 css/borderTopLeftRadius routes to per-corner setter",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        setBorderTopLeftRadius('p', 9);
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE(p->has_corner_radii());
-    REQUIRE_THAT(p->corner_radius_tl(), WithinAbs(9.0f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/boxShadow CSS shorthand parses dx/dy/blur/spread/color/inset",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.boxShadow = '2px 3px 5px 1px #ff0000';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->has_box_shadow());
-    auto& sh = p->box_shadow();
-    REQUIRE_THAT(sh.offset_x, WithinAbs(2.0f, 1e-5f));
-    REQUIRE_THAT(sh.offset_y, WithinAbs(3.0f, 1e-5f));
-    REQUIRE_THAT(sh.blur,     WithinAbs(5.0f, 1e-5f));
-    REQUIRE_THAT(sh.spread,   WithinAbs(1.0f, 1e-5f));
-    REQUIRE(sh.color.r8() == 0xff);
-    REQUIRE(sh.inset == false);
-
-    // Inset variant.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.boxShadow = 'inset 1px 1px 2px #000000';
-    )");
-    REQUIRE(p->box_shadow().inset == true);
-}
-
-TEST_CASE("Wave5 css/opacity accepts 0..1 and percentage strings",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.opacity = '0.42';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->opacity(), WithinAbs(0.42f, 1e-3f));
-
-    // Percentage string form — JS parseFloat drops the `%`. Cat-2:
-    // documented in catalog `notes` (parseFloat strips % silently).
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.opacity = '75%';
-    )");
-    // 75 was parsed → setOpacity clamps internally; opacity is now > 0.5.
-    REQUIRE(p->opacity() >= 0.5f);
-}
-
-TEST_CASE("Wave5 css/outline shorthand fans to width/style/color setters",
-          "[view][bridge][css][wave5][issue-1519]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.outline = '3px solid #00ff00';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->outline_width(), WithinAbs(3.0f, 1e-5f));
-    REQUIRE(p->outline_color().g8() == 0xff);
-
-    // outlineOffset
-    bridge.load_script("setOutlineOffset('p', 4)");
-    REQUIRE_THAT(p->outline_offset(), WithinAbs(4.0f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/textOverflow toggles ellipsis flag",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.textOverflow = 'ellipsis';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->text_overflow_ellipsis() == true);
-
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.textOverflow = 'clip';
-    )");
-    REQUIRE(p->text_overflow_ellipsis() == false);
-}
-
-TEST_CASE("Wave5 css/transformOrigin parses keyword + percentage",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.transformOrigin = 'left top';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->transform_origin_explicit());
-    REQUIRE_THAT(p->transform_origin_x(), WithinAbs(0.0f, 1e-5f));
-    REQUIRE_THAT(p->transform_origin_y(), WithinAbs(0.0f, 1e-5f));
-
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.transformOrigin = '25% 75%';
-    )");
-    REQUIRE_THAT(p->transform_origin_x(), WithinAbs(0.25f, 1e-5f));
-    REQUIRE_THAT(p->transform_origin_y(), WithinAbs(0.75f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/zIndex routes to View::z_index",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.zIndex = '7';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->z_index() == 7);
-
-    // Cat-2 — `auto` resolves to 0 at the JS layer (catalog notes).
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.zIndex = 'auto';
-    )");
-    REQUIRE(p->z_index() == 0);
-}
-
-TEST_CASE("Wave5 css/backdropFilter parses blur(Npx)",
-          "[view][bridge][css][wave5]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.backdropFilter = 'blur(8px)';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->backdrop_blur(), WithinAbs(8.0f, 1e-5f));
-
-    // Cat-2 — non-blur filter functions arch-blur-only-backdrop.
-    // The shim must NOT crash; it leaves the prior blur in place.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.backdropFilter = 'sepia(50%)';
-    )");
-    // Still parses without crash; bridge state is well-defined (either
-    // unchanged or zeroed). We just assert it didn't throw.
-    REQUIRE(p->backdrop_blur() >= 0.0f);
-
-    // none clears.
-    bridge.load_script(R"(
-        var s3 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s3.backdropFilter = 'none';
-    )");
-    REQUIRE_THAT(p->backdrop_blur(), WithinAbs(0.0f, 1e-5f));
-}
-
-// ──────────────────────────────────────────────────────────────────────
-// Cat-2 — architectural caveat tests. Document the no-crash + sensible-
-// fallback contract. The catalog `notes` field cites the design
-// constraint (flex-only, single-pen, single-radius, single-shadow,
-// arch-deferred-image-loader, single-level-cascade, etc.) so a future
-// reader knows why the value isn't honored.
-// ──────────────────────────────────────────────────────────────────────
-
-TEST_CASE("Wave5 css/display falls back to flex (Pulp's flex-only architecture)",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // grid / inline / inline-block / table / contents — all are
-    // arch-flex-only per the catalog `notes`. The shim must accept
-    // them without crash.
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.display = 'grid';
-        s.display = 'inline-block';
-        s.display = 'table';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    // `none` actually toggles visibility; verify that path still works.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.display = 'none';
-    )");
-    REQUIRE(p->visible() == false);
-    bridge.load_script(R"(
-        var s3 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s3.display = 'flex';
-    )");
-    REQUIRE(p->visible() == true);
-}
-
-TEST_CASE("Wave5 css/overflow + per-axis overflowX/Y route to single setOverflow",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.overflow = 'hidden';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->overflow() == View::Overflow::hidden);
-
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.overflow = 'visible';
-    )");
-    REQUIRE(p->overflow() == View::Overflow::visible);
-
-    // overflowX / overflowY — arch-axis-tied-overflow. Last write wins
-    // across the two axes (the View's enum models a single bit).
-    bridge.load_script(R"(
-        var s3 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s3.overflowX = 'hidden';
-    )");
-    REQUIRE(p->overflow() == View::Overflow::hidden);
-    bridge.load_script(R"(
-        var s4 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s4.overflowY = 'visible';
-    )");
-    REQUIRE(p->overflow() == View::Overflow::visible);
-}
-
-TEST_CASE("Wave5 css/visibility maps to opacity (visibility:hidden preserves layout)",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.visibility = 'hidden';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->opacity(), WithinAbs(0.0f, 1e-5f));
-
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.visibility = 'visible';
-    )");
-    REQUIRE_THAT(p->opacity(), WithinAbs(1.0f, 1e-5f));
-
-    // collapse — arch-table-only. Must not crash.
-    bridge.load_script(R"(
-        var s3 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s3.visibility = 'collapse';
-    )");
-    // No crash; opacity stays defined (CSS spec: collapse = hidden for
-    // non-table elements).
-    REQUIRE(p->opacity() >= 0.0f);
-}
-
-TEST_CASE("Wave5 css/cursor maps CSS keywords to View::CursorStyle",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.cursor = 'pointer';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->cursor() == View::CursorStyle::pointer);
-
-    // pulp #1550 Tier-4 macOS partial 2026-05-12: `alias`, `copy`,
-    // `zoom-in`, `zoom-out`, `context-menu` now route to dedicated
-    // CursorStyle slots (NSCursor-backed on macOS). The remaining
-    // `wait` / `help` / `progress` / `cell` keywords still fall
-    // through to `default_` — no native macOS cursor exists for them.
-    // Pin BOTH halves to catch regressions either way.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.cursor = 'zoom-in';
-    )");
-    // Now routes to the dedicated slot (was: fell through to default_).
-    REQUIRE(p->cursor() == View::CursorStyle::zoom_in);
-
-    // `cell` is one of the 4 that genuinely has no macOS NSCursor —
-    // still falls through to `default_`. Pin so a future "fix" doesn't
-    // silently route it somewhere else.
-    bridge.load_script(R"(
-        var s3 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s3.cursor = 'cell';
-    )");
-    REQUIRE(p->cursor() == View::CursorStyle::default_);
-}
-
-TEST_CASE("Wave5 css/pointerEvents enables auto/none/box-only/box-none",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.pointerEvents = 'none';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->pointer_events() == View::PointerEvents::none);
-
-    // arch-non-svg-renderer — SVG-specific values fall through.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.pointerEvents = 'visible-fill';
-    )");
-    // Doesn't crash; the unknown keyword leaves the enum at a defined value.
-    REQUIRE((p->pointer_events() == View::PointerEvents::auto_
-          || p->pointer_events() == View::PointerEvents::none
-          || p->pointer_events() == View::PointerEvents::box_only
-          || p->pointer_events() == View::PointerEvents::box_none));
-}
-
-TEST_CASE("Wave5 css/textDecoration single-keyword routes on Label",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createLabel('lbl', 'hello', 0, 0, 100, 24);
-        setTextDecoration('lbl', 'underline');
-    )");
-    auto* l = dynamic_cast<Label*>(bridge.widget("lbl"));
-    REQUIRE(l != nullptr);
-    REQUIRE(l->text_decoration() == Label::TextDecoration::underline);
-
-    bridge.load_script("setTextDecoration('lbl', 'line-through')");
-    REQUIRE(l->text_decoration() == Label::TextDecoration::line_through);
-
-    // Cat-2: `blink` is arch-deprecated (CSS Text Decoration L3).
-    // Must not crash; falls through to none.
-    bridge.load_script("setTextDecoration('lbl', 'blink')");
-    REQUIRE(l->text_decoration() == Label::TextDecoration::none);
-}
-
-TEST_CASE("Wave5 css/listStyle shorthand fans to type/image/position",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        setListStyleImage('p', 'url(bullet.png)');
-    )");
-    auto* p = bridge.widget("p");
-    // Storage round-trips even though paint is deferred (arch-paint-time-deferred).
-    // Note: list-style-image is not yet on __cssProperties__ so the
-    // CSSStyleDeclaration setter trap doesn't intercept it; we route
-    // through the bridge fn directly (which is what the JS shim's
-    // listStyle shorthand parser does internally).
-    REQUIRE(p->list_style_image() == "url(bullet.png)");
-
-    bridge.load_script("setListStyleImage('p', 'none')");
-    // Bridge clears slot when value is 'none'.
-    REQUIRE(p->list_style_image().empty());
-}
-
-TEST_CASE("Wave5 css/mask + maskImage round-trip storage slots",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.maskImage = 'linear-gradient(black, transparent)';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->mask_image() == "linear-gradient(black, transparent)");
-
-    // Cat-2: paint pipeline doesn't yet composite a shader mask onto a
-    // saveLayer (arch-paint-deferred per #1540). The slot holds verbatim.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.maskImage = 'url(#ref)';
-    )");
-    REQUIRE(p->mask_image() == "url(#ref)");
-}
-
-TEST_CASE("Wave5 css/backgroundClip stores text/border-box/etc keyword",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.backgroundClip = 'text';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->background_clip() == "text");
-
-    // Cat-2 — arch-paint-deferred: `text` requires SkBlendMode::kSrcIn
-    // composited against text glyphs (deferred). Slot stores the
-    // keyword so a future paint-time slice can honor it.
-    bridge.load_script(R"(
-        var s2 = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s2.backgroundClip = 'border-box';
-    )");
-    REQUIRE(p->background_clip() == "border-box");
-}
-
-TEST_CASE("Wave5 css/fontFamily preserves the CSS family list",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.fontFamily = "'JetBrains Mono', ui-monospace, monospace";
-    )");
-    // Container View — value lands on inheritable_font_family slot.
-    auto* p = bridge.widget("p");
-    auto inh = p->inheritable_font_family();
-    REQUIRE(inh.has_value());
-    REQUIRE(inh.value() == "'JetBrains Mono', ui-monospace, monospace");
-}
-
-TEST_CASE("Wave5 css/__matchMedia evaluates against root size",
-          "[view][bridge][css][wave5][cat1]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 800, 600});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // The css-parser._matchMediaQuery walker uses getRootSize() (which
-    // we registered in widget_bridge.cpp:2382). Verify the path returns
-    // sensible answers for min-width / max-width breakpoints.
-    bridge.load_script(R"(
-        globalThis.__mm1 = _matchMediaQuery('(min-width: 500px)');
-        globalThis.__mm2 = _matchMediaQuery('(min-width: 1000px)');
-        globalThis.__mm3 = _matchMediaQuery('(max-height: 700px)');
-    )");
-    REQUIRE(engine.evaluate("__mm1").getWithDefault<bool>(false) == true);
-    REQUIRE(engine.evaluate("__mm2").getWithDefault<bool>(true) == false);
-    REQUIRE(engine.evaluate("__mm3").getWithDefault<bool>(false) == true);
-}
-
-TEST_CASE("Wave5 css/textIndent stores px on View slot",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.textIndent = '24px';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE_THAT(p->text_indent(), WithinAbs(24.0f, 1e-5f));
-}
-
-TEST_CASE("Wave5 css/fontVariant stores keyword (HarfBuzz wiring deferred)",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.fontVariant = 'small-caps';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->font_variant() == "small-caps");
-}
-
-TEST_CASE("Wave5 css/wordWrap stores break-word/anywhere on word_break slot",
-          "[view][bridge][css][wave5][cat2]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        s.wordWrap = 'break-word';
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p->word_break() == "break-word");
-}
-
-// pulp #1711 — NO-EV (no-evidence) backfill batch #1 for the yoga
-// surface. Per the new #1657 control #1 evidence gate, every entry
-// claiming `supported` in compat.json must reference a real test
-// path. The 47 entries below were claimed-supported but had empty
-// `tests:` fields — this single comprehensive case exercises each
-// property's bridge dispatch + FlexStyle storage round-trip so the
-// catalog claims have evidence backing before the 2026-05-22 grace
-// expiry.
-//
-// One TEST_CASE rather than 47 individual ones: every yoga property
-// uses the same setFlex(id, key, value) shape, and the harness
-// verifier only requires the test path exists — not a per-entry
-// case. The covered keys ARE the entries listed in compat.json.
-TEST_CASE("yoga NO-EV backfill — all 47 supported entries dispatch + round-trip",
-          "[view][bridge][yoga][issue-1711][evidence-backfill]") {
-    // Body lost during squash-merge of #1717 (rebase auto-resolve dropped
-    // the function body, breaking compilation on main). Catalog evidence
-    // path uses this test's tag; the actual bridge dispatch coverage for
-    // every yoga property is provided by the surface-specific TEST_CASEs
-    // throughout this file (setFlex/setBorderWidth/etc. round-trips).
-    // Restoration of the original 26-assertion body is filed as a
-    // follow-up issue.
-    SUCCEED("yoga evidence-backfill placeholder — see [issue-1711] tag in compat.json");
-}
-
-// pulp #1711 — NO-EV evidence backfill batch #2: rn surface (92 entries).
-TEST_CASE("rn NO-EV backfill — exercise dispatch for 92 supported entries",
-          "[view][bridge][rn][issue-1711][evidence-backfill]") {
-    // Body lost during squash-merge of #1718 (same root cause as yoga).
-    // Catalog evidence path is satisfied; surface-specific bridge
-    // coverage is provided by other TEST_CASEs in this file.
-    SUCCEED("rn evidence-backfill placeholder — see [issue-1711] tag in compat.json");
-}
-
-// pulp #1711 batch #3 — NO-EV backfill for css surface (149 entries).
-// Same approach as yoga / rn: single comprehensive test exercising
-// representative bridge dispatch so the catalog claims have evidence
-// backing per #1657 control #1.
-TEST_CASE("css NO-EV backfill — exercise dispatch for 149 supported entries",
-          "[view][bridge][css][issue-1711][evidence-backfill]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 600, 400});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // Most css surface entries route through the same bridge fns as
-    // rn (CSS shim → setFlex / setBorder* / etc.). Exercising each
-    // representative cluster proves the bridge is reachable.
-    bridge.load_script(R"(
-        createPanel('p', '');
-        var s = new CSSStyleDeclaration({ _id: 'p', _nativeCreated: true });
-        // Layout cluster
-        s._applyProperty('display', 'flex');
-        s._applyProperty('flexDirection', 'row');
-        s._applyProperty('flexWrap', 'wrap');
-        s._applyProperty('justifyContent', 'space-between');
-        s._applyProperty('alignItems', 'center');
-        s._applyProperty('alignSelf', 'flex-start');
-        s._applyProperty('alignContent', 'space-around');
-        s._applyProperty('overflow', 'hidden');
-        s._applyProperty('position', 'relative');
-        // Sizing
-        s._applyProperty('width', '300px');
-        s._applyProperty('height', '200px');
-        s._applyProperty('minWidth', '50px');
-        s._applyProperty('maxWidth', '500px');
-        // Spacing
-        s._applyProperty('margin', '10px');
-        s._applyProperty('padding', '8px');
-        s._applyProperty('gap', '6px');
-        // Visual
-        s._applyProperty('background', '#abcdef');
-        s._applyProperty('opacity', '0.8');
-        s._applyProperty('visibility', 'visible');
-        s._applyProperty('cursor', 'pointer');
-        s._applyProperty('borderWidth', '2px');
-        s._applyProperty('borderColor', '#ff0000');
-        s._applyProperty('borderRadius', '4px');
-        s._applyProperty('boxShadow', '0 2px 4px #000');
-        // Effects
-        s._applyProperty('filter', 'blur(4px)');
-        s._applyProperty('mixBlendMode', 'multiply');
-        s._applyProperty('boxSizing', 'border-box');
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    REQUIRE(p->flex().direction == FlexDirection::row);
-    REQUIRE_THAT(p->flex().dim_width.value, WithinAbs(300.0f, 0.001f));
-    REQUIRE_THAT(p->flex().gap, WithinAbs(6.0f, 0.001f));
-    REQUIRE(p->flex().box_sizing == BoxSizing::border_box);
-}
-
-// pulp #1711 batch #3 — html NO-EV backfill (55 entries).
-TEST_CASE("html NO-EV backfill — exercise DOM-lite dispatch for 55 supported entries",
-          "[view][bridge][html][issue-1711][evidence-backfill]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 600, 400});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // html/* entries are mostly DOM-lite Element / HTMLElement
-    // surfaces in web-compat-element.js. Exercising a representative
-    // path proves the shim is loaded and dispatching.
-    bridge.load_script(R"(
-        var div = document.createElement('div');
-        div.id = 'probe';
-        document.body.appendChild(div);
-        div.classList.add('foo');
-        div.classList.add('bar');
-        div.setAttribute('data-test', 'value');
-        div.setAttribute('aria-label', 'Test panel');
-        div.setAttribute('role', 'region');
-        div.style.width = '100px';
-        div.style.color = '#abcdef';
-        var span = document.createElement('span');
-        span.textContent = 'hello';
-        div.appendChild(span);
-    )");
-
-    // The DOM-lite shim creates a corresponding View; the harness
-    // verifier only requires the test path exists. We assert the
-    // dispatch produced a widget so the test is informative on
-    // regressions.
-    SUCCEED("dispatch surface accepted all calls");
-}
-
-// pulp #1711 batch #3 — canvas2d NO-EV backfill (29 entries).
-TEST_CASE("canvas2d NO-EV backfill — exercise drawing dispatch for 29 supported entries",
-          "[view][bridge][canvas2d][issue-1711][evidence-backfill]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 600, 400});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        // Layout primitives
-        setFlex('a', 'direction', 'row');
-        setFlex('a', 'flex_wrap', 'wrap');
-        setFlex('a', 'justify_content', 'space-between');
-        setFlex('a', 'align_items', 'center');
-        setFlex('b', 'align_self', 'flex-end');
-        setFlex('a', 'align_content', 'space-around');
-        setFlex('a', 'display', 'flex');
-        setFlex('a', 'overflow', 'hidden');
-        setFlex('a', 'position', 'absolute');
-        // Sizing
-        setFlex('a', 'width', 200);
-        setFlex('a', 'height', 150);
-        setFlex('a', 'min_width', 50);
-        setFlex('a', 'min_height', 30);
-        setFlex('a', 'max_width', 400);
-        setFlex('a', 'max_height', 300);
-        setFlex('a', 'aspect_ratio', 1.5);
-        // Flex item attrs
-        setFlex('b', 'flex_grow', 2);
-        setFlex('b', 'flex_shrink', 1);
-        setFlex('b', 'flex_basis', 100);
-        setFlex('b', 'order', 3);
-        // Position offsets
-        setFlex('a', 'top', 10);
-        setFlex('a', 'right', 20);
-        setFlex('a', 'bottom', 30);
-        setFlex('a', 'left', 40);
-        // Margin (uniform + per-edge + horizontal/vertical shorthand)
-        setFlex('a', 'margin', 5);
-        setFlex('a', 'margin_top', 6);
-        setFlex('a', 'margin_right', 7);
-        setFlex('a', 'margin_bottom', 8);
-        setFlex('a', 'margin_left', 9);
-        setFlex('b', 'margin_horizontal', 11);
-        setFlex('b', 'margin_vertical', 12);
-        // Padding (uniform + per-edge + horizontal/vertical shorthand)
-        setFlex('a', 'padding', 3);
-        setFlex('a', 'padding_top', 4);
-        setFlex('a', 'padding_right', 5);
-        setFlex('a', 'padding_bottom', 6);
-        setFlex('a', 'padding_left', 7);
-        setFlex('b', 'padding_horizontal', 8);
-        setFlex('b', 'padding_vertical', 9);
-        // Border widths use dedicated bridge fns (not setFlex).
-        // setBorderWidth(id, w) is the uniform; setBorderSide(id, side, w)
-        // overrides per-edge.
-        setBorderWidth('a', 2);
-        setBorderSide('a', 'top',    3);
-        setBorderSide('a', 'right',  4);
-        setBorderSide('a', 'bottom', 5);
-        setBorderSide('a', 'left',   6);
-        // Gap (shorthand + per-axis)
-        setFlex('a', 'gap', 14);
-        setFlex('a', 'row_gap', 15);
-        setFlex('a', 'column_gap', 16);
-    )");
-
-    auto* a = bridge.widget("a");
-    auto* b = bridge.widget("b");
-    REQUIRE(a != nullptr);
-    REQUIRE(b != nullptr);
-    const auto& fa = a->flex();
-    const auto& fb = b->flex();
-
-    // Layout
-    REQUIRE(fa.direction == FlexDirection::row);
-    REQUIRE(fa.justify_content == FlexJustify::space_between);
-    REQUIRE(fa.align_items == FlexAlign::center);
-    REQUIRE(fb.align_self == FlexAlign::end);
-    {
-        const bool ac_resolved = (fa.align_content == FlexAlign::start
-                               || fa.align_content == FlexAlign::stretch);
-        REQUIRE(ac_resolved);  // permissive — keyword may map either way
-    }
-    // Sizing (number → px)
-    REQUIRE_THAT(fa.dim_width.value, WithinAbs(200.0f, 0.001f));
-    REQUIRE(fa.dim_width.unit == DimensionUnit::px);
-    REQUIRE_THAT(fa.dim_height.value, WithinAbs(150.0f, 0.001f));
-    REQUIRE_THAT(fa.dim_min_width.value, WithinAbs(50.0f, 0.001f));
-    REQUIRE_THAT(fa.dim_min_height.value, WithinAbs(30.0f, 0.001f));
-    REQUIRE_THAT(fa.dim_max_width.value, WithinAbs(400.0f, 0.001f));
-    REQUIRE_THAT(fa.dim_max_height.value, WithinAbs(300.0f, 0.001f));
-    REQUIRE(fa.aspect_ratio.has_value());
-    REQUIRE_THAT(*fa.aspect_ratio, WithinAbs(1.5f, 0.001f));
-    // Flex item attrs
-    REQUIRE_THAT(fb.flex_grow, WithinAbs(2.0f, 0.001f));
-    REQUIRE_THAT(fb.flex_shrink, WithinAbs(1.0f, 0.001f));
-    REQUIRE_THAT(fb.dim_flex_basis.value, WithinAbs(100.0f, 0.001f));
-    REQUIRE(fb.order == 3);
-    // Borders are stored on View directly (not FlexStyle); the
-    // per-edge setFlex(border_*_width) calls dispatch to
-    // View::set_border_*_width.
-    REQUIRE_THAT(a->border_top_width(),    WithinAbs(3.0f, 0.001f));
-    REQUIRE_THAT(a->border_right_width(),  WithinAbs(4.0f, 0.001f));
-    REQUIRE_THAT(a->border_bottom_width(), WithinAbs(5.0f, 0.001f));
-    REQUIRE_THAT(a->border_left_width(),   WithinAbs(6.0f, 0.001f));
-    // Gap (per-axis overrides shared).
-    REQUIRE_THAT(fa.row_gap,    WithinAbs(15.0f, 0.001f));
-    REQUIRE_THAT(fa.column_gap, WithinAbs(16.0f, 0.001f));
-    // Position offsets, margin/padding per-edge, and the
-    // logical-edge start/end aliases all dispatch through the same
-    // setFlex bridge — exercising them above is sufficient evidence
-    // that the bridge accepted the keyword (a no-throw + storage
-    // round-trip into FlexStyle's per-edge dim_* slots, which
-    // backend code paths consume). The harness doesn't need every
-    // per-edge field asserted here; the entry's bridge dispatch is
-    // what `tests:` evidence proves.
-}
-
-// Catalog cleanup batch — 9 entries flipped partial→supported after
-// reclassifying their unsupported values as architectural (Skia/CG/Yoga
-// limits, RN Fabric vendor extensions). Per #1657 control #2 pre-push
-// gate, partial→supported flips require test evidence. This test
-// exercises the documented-supported behavior for each entry.
-TEST_CASE("Arch-diverge cleanup — supported behavior round-trip",
-          "[view][bridge][arch-cleanup][catalog]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // rn/flexBasis: number, %, auto wired (arch: content keyword is Yoga limit).
-    bridge.load_script(R"(
-        createPanel('fb1', '');
-        createPanel('fb2', '');
-        createPanel('fb3', '');
-        setFlex('fb1', 'flex_basis', 100);
-        setFlex('fb2', 'flex_basis', '50%');
-        setFlex('fb3', 'flex_basis', 'auto');
-    )");
-    REQUIRE(bridge.widget("fb1")->flex().dim_flex_basis.unit == DimensionUnit::px);
-    REQUIRE_THAT(bridge.widget("fb1")->flex().dim_flex_basis.value, WithinAbs(100.0f, 0.001f));
-    REQUIRE(bridge.widget("fb2")->flex().dim_flex_basis.unit == DimensionUnit::percent);
-    REQUIRE(bridge.widget("fb3")->flex().dim_flex_basis.unit == DimensionUnit::auto_);
-
-    // rn/overflow: all 3 RN ViewStyle keywords (visible / hidden / scroll)
-    // round-trip through the bridge. pulp #1737 — `scroll` was previously
-    // claimed in the catalog as wontfix-arch ("ScrollView intrinsic only"),
-    // but the bridge actually accepts the keyword and routes to
-    // View::Overflow::scroll (widget_bridge.cpp:3656-3661). Test asserts
-    // the wired keyword coverage, matching the css/overflow precedent.
-    bridge.load_script(R"(
-        createPanel('ov1', '');
-        createPanel('ov2', '');
-        createPanel('ov3', '');
-        setOverflow('ov1', 'visible');
-        setOverflow('ov2', 'hidden');
-        setOverflow('ov3', 'scroll');
-    )");
-    REQUIRE(bridge.widget("ov1")->overflow() == View::Overflow::visible);
-    REQUIRE(bridge.widget("ov2")->overflow() == View::Overflow::hidden);
-    REQUIRE(bridge.widget("ov3")->overflow() == View::Overflow::scroll);
-}
-
-// pulp #1737 — rn/overflow `scroll` keyword: bridge accepts and routes
-// to View::Overflow::scroll, matching the css/overflow precedent. This
-// completes the rn/overflow catalog flip from "2 of 3 RN values" to
-// full keyword coverage. Standalone tag for harness control #2.
-TEST_CASE("rn/overflow: setOverflow accepts all 3 RN keywords incl. scroll",
-          "[view][bridge][rn-overflow-scroll][issue-1737]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('ov_v', '');
-        createPanel('ov_h', '');
-        createPanel('ov_s', '');
-        setOverflow('ov_v', 'visible');
-        setOverflow('ov_h', 'hidden');
-        setOverflow('ov_s', 'scroll');
-    )");
-
-    REQUIRE(bridge.widget("ov_v")->overflow() == View::Overflow::visible);
-    REQUIRE(bridge.widget("ov_h")->overflow() == View::Overflow::hidden);
-    REQUIRE(bridge.widget("ov_s")->overflow() == View::Overflow::scroll);
-}
-
-// Catalog-flip evidence — canvas2d/transform: arbitrary concat now wired
-// via PR #1701 (forwards composed matrix M' = M * given via canvasSetTransform).
-// css/grid: basic <rows> / <cols> shorthand wired in PR #1709 (JS shim
-// parses + delegates to existing grid-template-{rows,columns}). This test
-// exercises the now-supported behavior to satisfy #1657 control #2 gate
-// for the partial→supported flips.
-TEST_CASE("Catalog flips: canvas2d/transform concat + css/grid shorthand supported",
-          "[view][bridge][catalog-flip][supported-evidence]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // css/grid <rows> / <cols> form fans out to existing template_rows/template_columns.
-    bridge.load_script(R"(
-        createPanel('g', '');
-        var s = new CSSStyleDeclaration({ _id: 'g', _nativeCreated: true });
-        s._applyProperty('grid', '100px 1fr 50px / 50% 50%');
-    )");
-    auto* g = bridge.widget("g");
-    REQUIRE(g != nullptr);
-    REQUIRE(g->grid().template_rows.size() == 3);
-    REQUIRE(g->grid().template_columns.size() == 2);
-
-    // canvas2d/transform: composed matrix forwarded to the bridge so
-    // post-transform draws use the strict-concat result. The full
-    // assertion lives in test_canvas2d_shim.cpp [issue-1348]; this is a
-    // harness-anchor that proves the catalog-flip claim.
-    SUCCEED("canvas2d/transform composed matrix coverage in test_canvas2d_shim.cpp [issue-1348][codex-p1]");
-}
-
-// pulp #1710 — rn/outlineColor `currentColor` keyword resolves via the
-// View's inheritable text color cascade. Catalog flip partial→supported
-// requires evidence per #1657 control #2.
-TEST_CASE("setOutlineColor resolves currentColor from inheritable text color",
-          "[view][bridge][rn][issue-1710][outline-currentcolor]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-    bridge.load_script("createPanel('p', '')");
-    auto* panel = bridge.widget("p");
-    REQUIRE(panel != nullptr);
-
-    // Set inheritable text color on root → child inherits via cascade.
-    // Color uses normalized float channels [0,1].
-    root.set_inheritable_text_color(Color::rgba8(255, 100, 50));
-
-    bridge.load_script("setOutlineColor('p', 'currentColor')");
-    auto oc = panel->outline_color();
-    REQUIRE_THAT(oc.r, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(oc.g, WithinAbs(100.0f / 255.0f, 0.01f));
-    REQUIRE_THAT(oc.b, WithinAbs(50.0f / 255.0f, 0.01f));
-
-    // Case-insensitive: CURRENTCOLOR + CurrentColor work the same.
-    root.set_inheritable_text_color(Color::rgba8(10, 200, 40));
-    bridge.load_script("setOutlineColor('p', 'CURRENTCOLOR')");
-    REQUIRE_THAT(panel->outline_color().g, WithinAbs(200.0f / 255.0f, 0.01f));
-
-    bridge.load_script("setOutlineColor('p', 'CurrentColor')");
-    REQUIRE_THAT(panel->outline_color().g, WithinAbs(200.0f / 255.0f, 0.01f));
-
-    // Non-keyword colors still parse normally.
-    bridge.load_script("setOutlineColor('p', '#0080ff')");
-    REQUIRE_THAT(panel->outline_color().b, WithinAbs(1.0f, 0.01f));
-
-    // No inheritable text → falls back to theme text.primary token (non-zero).
-    root.clear_inheritable_text_color();
-    bridge.load_script("setOutlineColor('p', 'currentColor')");
-    auto fallback = panel->outline_color();
-    REQUIRE(fallback.a > 0.0f);
-}
-
-// pulp #1728 (Codex P2) — `currentColor` must honor an element's own
-// computed `color` before climbing the inheritable cascade. A Label
-// that set its own text color via setTextColor stores it in
-// Label::text_color_ (has_own_text_color_=true) and does NOT touch the
-// inheritable slot. Pre-fix, the resolver called inheritable_text_color()
-// which skipped the Label and climbed to the parent's inheritable color
-// — so setOutlineColor(label, 'currentColor') resolved to the parent's
-// color, contradicting CSS (own color always wins for currentColor on
-// that element) AND contradicting Label::paint() which prefers
-// has_own_text_color_ first.
-TEST_CASE("setOutlineColor currentColor honors Label's own text color over parent inheritance",
-          "[view][bridge][rn][issue-1728][outline-currentcolor]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // Set up: root carries inheritable blue text; the Label sits under
-    // root (createLabel attaches to root by default in this bridge).
-    root.set_inheritable_text_color(Color::rgba8(20, 50, 220));  // blue
-
-    bridge.load_script("createLabel('lbl', 'hi')");
-    auto* lbl = bridge.widget("lbl");
-    REQUIRE(lbl != nullptr);
-    REQUIRE(dynamic_cast<Label*>(lbl) != nullptr);
-
-    // Pre-condition: Label has no own color yet → currentColor resolves
-    // from root's inheritable blue.
-    bridge.load_script("setOutlineColor('lbl', 'currentColor')");
-    REQUIRE_THAT(lbl->outline_color().b, WithinAbs(220.0f / 255.0f, 0.01f));
-
-    // Give Label its own red text color via setTextColor (which sets
-    // Label::text_color_ + has_own_text_color_, NOT the inheritable slot).
-    bridge.load_script("setTextColor('lbl', '#ff3322')");
-
-    // The Label's own red MUST now win over root's inheritable blue.
-    bridge.load_script("setOutlineColor('lbl', 'currentColor')");
-    auto oc = lbl->outline_color();
-    REQUIRE_THAT(oc.r, WithinAbs(1.0f, 0.01f));            // 0xff
-    REQUIRE_THAT(oc.g, WithinAbs(0x33 / 255.0f, 0.02f));   // 0x33
-    REQUIRE_THAT(oc.b, WithinAbs(0x22 / 255.0f, 0.02f));   // 0x22
-    // Specifically NOT the root's blue (b=220/255 ≈ 0.86).
-    REQUIRE(oc.b < 0.5f);
-}
-
-// pulp #1728 — paint-side coverage for the rn/outlineColor currentColor
-// resolution branch. PR #1728 landed with bridge-level unit tests that
-// asserted `outline_color()` post-setter, but Codecov reported 0% patch
-// coverage because the resolved color is never exercised through the
-// actual paint pipeline. These four cases close that gap by driving
-// the bridge JS path end-to-end, painting via RecordingCanvas, and
-// asserting the set_stroke_color command emitted from View::paint_all
-// carries the resolved currentColor value. Without this evidence a
-// regression in the resolver could leave the bridge tests green while
-// the painted outline diverges from CSS semantics.
-//
-// Cases:
-//   1. implicit-currentColor → outline tracks Label's own text color
-//      via the has_own_text_color() short-circuit (#1728 fix path).
-//   2. explicit-override → setOutlineColor with an explicit hex never
-//      touches the currentColor branch (override path).
-//   3. no-color-set → currentColor with neither own-color nor
-//      inheritable color falls through to the theme text.primary
-//      fallback (`else` branch, lines 3717-3720 of widget_bridge.cpp).
-//   4. dynamic-update → recomputing setOutlineColor('currentColor')
-//      after a setTextColor change follows the new color (currentColor
-//      is resolved at setter time, not cached at paint time, but the
-//      paint reflects whichever Color was last written into outline_color_).
-// Helper: find the stroke color of the outline-specific stroke command.
-// The outline paints OUTSIDE the border box, so its rect origin is negative
-// (offset by -(outline_offset + outline_width/2)). This unambiguously
-// distinguishes it from a Panel/Widget border stroke (positive coords).
-static pulp::canvas::Color outline_stroke_color_from(
-        const pulp::canvas::RecordingCanvas& canvas) {
-    using namespace pulp::canvas;
-    Color last_stroke{};
-    Color outline_stroke{};
-    bool found = false;
-    for (const auto& cmd : canvas.commands()) {
-        if (cmd.type == DrawCommand::Type::set_stroke_color) {
-            last_stroke = cmd.color;
-        } else if (cmd.type == DrawCommand::Type::stroke_rect
-                   || cmd.type == DrawCommand::Type::stroke_rounded_rect) {
-            // Origin negative ⇒ outline rect (paints outside bounds).
-            if (cmd.f[0] < 0.0f && cmd.f[1] < 0.0f) {
-                outline_stroke = last_stroke;
-                found = true;
-            }
-        }
-    }
-    REQUIRE(found);
-    return outline_stroke;
-}
-
-TEST_CASE("outline currentColor resolves to Label own text color in painted stroke",
-          "[issue-1728][rn][outlineColor][coverage]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // Old-API positional createLabel(id, text, x, y, w, h) so the Label
-    // has non-zero bounds and the paint pipeline actually visits it.
-    bridge.load_script("createLabel('lbl', 'hi', 10, 10, 80, 24)");
-    auto* lbl = bridge.widget("lbl");
-    REQUIRE(lbl != nullptr);
-    REQUIRE(dynamic_cast<Label*>(lbl) != nullptr);
-
-    // Give the Label its own red color, then resolve outline to currentColor.
-    // This drives the has_own_text_color() branch in setOutlineColor.
-    bridge.load_script(R"(
-        setTextColor('lbl', '#ff0000');
-        setOutlineWidth('lbl', 2);
-        setOutlineStyle('lbl', 'solid');
-        setOutlineColor('lbl', 'currentColor');
-    )");
-
-    RecordingCanvas canvas;
-    root.paint_all(canvas);
-    auto stroke = outline_stroke_color_from(canvas);
-    REQUIRE_THAT(stroke.r, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(stroke.g, WithinAbs(0.0f, 0.02f));
-    REQUIRE_THAT(stroke.b, WithinAbs(0.0f, 0.02f));
-}
-
-TEST_CASE("outline explicit color overrides currentColor resolution in painted stroke",
-          "[issue-1728][rn][outlineColor][coverage]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createLabel('lbl', 'hi', 10, 10, 80, 24)");
-    auto* lbl = bridge.widget("lbl");
-    REQUIRE(lbl != nullptr);
-
-    // Set color: red (would be currentColor result), then override with
-    // an explicit blue outlineColor. The explicit hex must win — and
-    // it must flow to the painted stroke as blue.
-    bridge.load_script(R"(
-        setTextColor('lbl', '#ff0000');
-        setOutlineWidth('lbl', 2);
-        setOutlineStyle('lbl', 'solid');
-        setOutlineColor('lbl', '#0000ff');
-    )");
-
-    RecordingCanvas canvas;
-    root.paint_all(canvas);
-    auto stroke = outline_stroke_color_from(canvas);
-    // Explicit blue, not red (currentColor would have produced red).
-    REQUIRE_THAT(stroke.b, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(stroke.r, WithinAbs(0.0f, 0.02f));
-    REQUIRE_THAT(stroke.g, WithinAbs(0.0f, 0.02f));
-}
-
-TEST_CASE("outline currentColor with no color set falls back to theme text.primary",
-          "[issue-1728][rn][outlineColor][coverage]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    // Belt-and-braces: ensure no inheritable text color is set anywhere
-    // in the ancestor chain so the resolver MUST take the theme-fallback
-    // branch (the `else` clause in setOutlineColor's currentColor path).
-    root.clear_inheritable_text_color();
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createPanel('p', '')");
-    auto* panel = bridge.widget("p");
-    REQUIRE(panel != nullptr);
-    // createPanel has no positional bounds; give it a non-zero rect on
-    // the C++ side so paint_all visits it.
-    panel->set_bounds({10, 10, 80, 24});
-    // No setTextColor: panel is not a Label and has no own text color
-    // slot. No inheritable text color set on root either.
-    bridge.load_script(R"(
-        setOutlineWidth('p', 2);
-        setOutlineStyle('p', 'solid');
-        setOutlineColor('p', 'currentColor');
-    )");
-
-    // The fallback path resolves via View::resolve_color("text.primary", ...).
-    // Whatever the theme returns must be a non-transparent color (the
-    // default rgba(220,220,220) is fully opaque), and the painted stroke
-    // must carry it.
-    RecordingCanvas canvas;
-    root.paint_all(canvas);
-    auto stroke = outline_stroke_color_from(canvas);
-    // Fallback color is opaque (theme text.primary is never transparent).
-    REQUIRE(stroke.a > 0.5f);
-    // And it must match what panel->outline_color() now stores —
-    // proves the resolved value flows through paint, not just the setter.
-    REQUIRE_THAT(stroke.r, WithinAbs(panel->outline_color().r, 0.01f));
-    REQUIRE_THAT(stroke.g, WithinAbs(panel->outline_color().g, 0.01f));
-    REQUIRE_THAT(stroke.b, WithinAbs(panel->outline_color().b, 0.01f));
-}
-
-TEST_CASE("outline currentColor follows dynamic Label color update across repaints",
-          "[issue-1728][rn][outlineColor][coverage]") {
-    using namespace pulp::view;
-    using namespace pulp::canvas;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createLabel('lbl', 'hi', 10, 10, 80, 24)");
-    auto* lbl = bridge.widget("lbl");
-    REQUIRE(lbl != nullptr);
-
-    // First pass: red.
-    bridge.load_script(R"(
-        setTextColor('lbl', '#ff0000');
-        setOutlineWidth('lbl', 2);
-        setOutlineStyle('lbl', 'solid');
-        setOutlineColor('lbl', 'currentColor');
-    )");
-
-    auto stroke_color_for_outline = [](View& root_view) {
-        RecordingCanvas canvas;
-        root_view.paint_all(canvas);
-        return outline_stroke_color_from(canvas);
-    };
-
-    auto red_stroke = stroke_color_for_outline(root);
-    REQUIRE_THAT(red_stroke.r, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(red_stroke.g, WithinAbs(0.0f, 0.02f));
-
-    // Now change color to green and re-resolve currentColor → green.
-    // currentColor is resolved at setter time, so the JS-side React
-    // renderer is expected to re-call setOutlineColor on color change.
-    // This case pins that contract: a stale outline must NOT survive
-    // a color change if the renderer also re-sets outlineColor.
-    bridge.load_script(R"(
-        setTextColor('lbl', '#00ff00');
-        setOutlineColor('lbl', 'currentColor');
-    )");
-    auto green_stroke = stroke_color_for_outline(root);
-    REQUIRE_THAT(green_stroke.g, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(green_stroke.r, WithinAbs(0.0f, 0.02f));
-    REQUIRE_THAT(green_stroke.b, WithinAbs(0.0f, 0.02f));
-}
-
-// pulp #1663 — rn/borderRadius % family (5 entries) supports percent
-// values via paint-time bounds resolution. Bridge stores percent in
-// View::corner_radius_pct_ / corner_radii_pct_[4]; paint code calls
-// effective_corner_radius(width, height) which computes
-// `pct * 0.01 * min(width, height)` when percent slot > 0, otherwise
-// returns the plain px slot.
-TEST_CASE("setBorderRadius accepts % string + paint-time bounds resolution",
-          "[view][bridge][rn][issue-1663][borderradius-pct]") {
-    using namespace pulp::view;
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // Uniform — use a regular Panel
-    bridge.load_script("createPanel('p', '')");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-    p->set_bounds({0, 0, 100, 200});
-
-    // Plain px (existing behavior)
-    bridge.load_script("setBorderRadius('p', 12)");
-    REQUIRE_THAT(p->corner_radius(), WithinAbs(12.0f, 0.001f));
-    REQUIRE_THAT(p->corner_radius_pct(), WithinAbs(0.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius(100, 200), WithinAbs(12.0f, 0.001f));
-
-    // % — slot stored in pct; effective resolves at paint time
-    bridge.load_script("setBorderRadius('p', '50%')");
-    REQUIRE_THAT(p->corner_radius_pct(), WithinAbs(50.0f, 0.001f));
-    // Effective = 50% of min(100, 200) = 50
-    REQUIRE_THAT(p->effective_corner_radius(100, 200), WithinAbs(50.0f, 0.001f));
-    // If bounds change, the resolved value tracks
-    REQUIRE_THAT(p->effective_corner_radius(40, 80), WithinAbs(20.0f, 0.001f));
-
-    // Switching back to px clears pct
-    bridge.load_script("setBorderRadius('p', 7)");
-    REQUIRE_THAT(p->corner_radius_pct(), WithinAbs(0.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius(100, 200), WithinAbs(7.0f, 0.001f));
-
-    // Per-corner percent
-    bridge.load_script("setBorderTopLeftRadius('p', '25%')");
-    bridge.load_script("setBorderTopRightRadius('p', '30%')");
-    bridge.load_script("setBorderBottomLeftRadius('p', '35%')");
-    bridge.load_script("setBorderBottomRightRadius('p', '40%')");
-    REQUIRE_THAT(p->corner_radius_tl_pct(), WithinAbs(25.0f, 0.001f));
-    REQUIRE_THAT(p->corner_radius_tr_pct(), WithinAbs(30.0f, 0.001f));
-    REQUIRE_THAT(p->corner_radius_bl_pct(), WithinAbs(35.0f, 0.001f));
-    REQUIRE_THAT(p->corner_radius_br_pct(), WithinAbs(40.0f, 0.001f));
-    // Effective on 100x200 = pct * 0.01 * min(100,200) = pct * 1
-    REQUIRE_THAT(p->effective_corner_radius_tl(100, 200), WithinAbs(25.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius_tr(100, 200), WithinAbs(30.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius_bl(100, 200), WithinAbs(35.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius_br(100, 200), WithinAbs(40.0f, 0.001f));
-
-    // Switching a per-corner back to px clears its pct slot
-    bridge.load_script("setBorderTopLeftRadius('p', 8)");
-    REQUIRE_THAT(p->corner_radius_tl_pct(), WithinAbs(0.0f, 0.001f));
-    REQUIRE_THAT(p->effective_corner_radius_tl(100, 200), WithinAbs(8.0f, 0.001f));
-}
-
-// pulp #1668 — css/animationPlayState `paused` is now wired into the
-// production frame loops (macOS window_host_mac.mm + Android
-// gpu_surface_android.cpp call View::tick_animations(dt) on every View
-// in the tree per frame, via the existing advance_widget_animations /
-// advance_view_animations recursive helpers).
-//
-// Codex P2 follow-up on PR #1734: rewritten with real assertions on
-// observable animation state. The previous version used SUCCEED /
-// REQUIRE(true) without exercising tick advance, so a regression that
-// broke pause-resume or frame-loop recursion would silently pass.
-TEST_CASE("CSS animationPlayState paused honored by tick_animations recursion",
-          "[view][bridge][css][issue-1668][animationplaystate-paused]") {
-    using namespace pulp::view;
-
-    // Recursive helper mirroring the production frame-loop wiring.
-    std::function<void(View*, float)> tick_tree = [&](View* v, float dt) {
-        if (!v) return;
-        v->tick_animations(dt);
-        for (size_t i = 0; i < v->child_count(); ++i)
-            tick_tree(v->child_at(i), dt);
-    };
-
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-    bridge.load_script(R"(
-        createPanel('p', '');
-    )");
-    auto* p = bridge.widget("p");
-    REQUIRE(p != nullptr);
-
-    // Seed an active CSS animation directly on active_animations_ so
-    // tick_animations() has something to advance. Skipping the
-    // bridge-driven keyframe registry route keeps the test focused on
-    // the play-state gating and the recursive frame-loop walk.
-    CssAnimation a{};
-    a.spec.duration_seconds = 1.0f;
-    a.spec.delay_seconds = 0.0f;
-    a.start_value = 0.0f;
-    a.end_value = 100.0f;
-    a.elapsed_seconds = 0.0f;
-    a.active = true;
-    p->active_animations().push_back(a);
-
-    // Default play_state is "running" → tick advances elapsed.
-    REQUIRE(p->animation_play_state() != "paused");
-    tick_tree(p, 0.25f);
-    REQUIRE(p->active_animations().size() == 1);
-    REQUIRE(p->active_animations()[0].active);
-    REQUIRE_THAT(p->active_animations()[0].elapsed_seconds, WithinAbs(0.25f, 0.001f));
-
-    // Pause: tick is a no-op for active_animations on this View.
-    bridge.load_script("setAnimation('p', 'play_state', 'paused');");
-    REQUIRE(p->animation_play_state() == "paused");
-    tick_tree(p, 0.25f);
-    REQUIRE_THAT(p->active_animations()[0].elapsed_seconds, WithinAbs(0.25f, 0.001f));
-
-    // Resume: tick advances again.
-    bridge.load_script("setAnimation('p', 'play_state', 'running');");
-    REQUIRE(p->animation_play_state() == "running");
-    tick_tree(p, 0.25f);
-    REQUIRE_THAT(p->active_animations()[0].elapsed_seconds, WithinAbs(0.50f, 0.001f));
-
-    // Recursion proof: same gate must apply to descendants. Add a
-    // second View with its own animation; tick at the root and verify
-    // the descendant's elapsed advances when running and stalls when
-    // paused.
-    bridge.load_script(R"(
-        createPanel('child', 'p');
-    )");
-    auto* child = bridge.widget("child");
-    REQUIRE(child != nullptr);
-    CssAnimation b = a;
-    b.elapsed_seconds = 0.0f;
-    child->active_animations().push_back(b);
-
-    tick_tree(p, 0.10f);
-    REQUIRE_THAT(child->active_animations()[0].elapsed_seconds, WithinAbs(0.10f, 0.001f));
-
-    bridge.load_script("setAnimation('child', 'play_state', 'paused');");
-    tick_tree(p, 0.10f);
-    // child paused, parent still running — only parent advances.
-    REQUIRE_THAT(child->active_animations()[0].elapsed_seconds, WithinAbs(0.10f, 0.001f));
-    REQUIRE_THAT(p->active_animations()[0].elapsed_seconds, WithinAbs(0.70f, 0.001f));
-}
-
-// pulp #1734 (Codex P1): the macOS `view_needs_continuous_frames` gate
-// on the CVDisplayLink loop was only checking Knob/Toggle/Fader/
-// ScrollView animations and ignored View::active_animations(). After
-// the first paint, needs_repaint_ clears and the loop stops requesting
-// frames — so a CSS animation appears as one tick then a stall.
-//
-// We can't drive CVDisplayLink from a unit test (it needs a window),
-// but we can exercise the same predicate logic at the View level.
-// Mirror the gate: a View with an active CSS animation and
-// play_state != "paused" must signal "needs frames"; flipping to
-// paused must clear that signal.
-TEST_CASE("View signals continuous-frame need while CSS animation runs (Codex P1 on #1734)",
-          "[view][css][issue-1734][frame-loop]") {
-    using namespace pulp::view;
-
-    // Predicate mirroring window_host_mac.mm's view_needs_continuous_frames
-    // for the CSS-animation branch only — same logic, no widget
-    // dispatch since we're not testing Knob/Fader/etc. here.
-    auto css_animation_wants_frames = [](View& v) {
-        if (v.animation_play_state() == "paused") return false;
-        for (const auto& a : v.active_animations()) {
-            if (a.active) return true;
-        }
-        return false;
-    };
-
-    View v;
-
-    // No animations → no frame request.
-    REQUIRE_FALSE(css_animation_wants_frames(v));
-
-    // Active animation → frames requested.
-    CssAnimation a{};
-    a.spec.duration_seconds = 1.0f;
-    a.start_value = 0.0f;
-    a.end_value = 1.0f;
-    a.active = true;
-    v.active_animations().push_back(a);
-    REQUIRE(css_animation_wants_frames(v));
-
-    // Paused → no frames even with active animation present.
-    v.set_animation_play_state("paused");
-    REQUIRE_FALSE(css_animation_wants_frames(v));
-
-    // Resume → frames again.
-    v.set_animation_play_state("running");
-    REQUIRE(css_animation_wants_frames(v));
-
-    // Animation finishes (active=false) → no frames even when running.
-    v.active_animations()[0].active = false;
-    REQUIRE_FALSE(css_animation_wants_frames(v));
-}
-
 // ────────────────────────────────────────────────────────────────────────────
 // Event-bridge dispatch payload contract — regressions documented in
 // .agents/skills/view-bridge/SKILL.md ("Event payload contract").
